toxTaskProvider: Re-use testenv parsing from run, move error checks there
Instead of having two code sections with the same purpose (one with a
more elaborate error handling), fuse them to one.
